# src/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/solvers": {
      "post": {
        "operationId": "SolversController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSolverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      },
      "get": {
        "operationId": "SolversController_getLegacyLeaderboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/leaderboard": {
      "get": {
        "description": "Returns the ranked solver list for a specific window. This endpoint is intended for recent-performance visibility and does not alter the legacy all-time leaderboard.",
        "operationId": "SolversController_getLeaderboard",
        "parameters": [
          {
            "name": "window",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "24h",
                "7d",
                "30d",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Windowed solver leaderboard",
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/eligible-intents": {
      "get": {
        "operationId": "SolversController_getEligibleIntents",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Filter by intent state",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "accepted",
                "filled",
                "cancelled",
                "expired",
                "slashed",
                "pending_open",
                "pending_accepted",
                "pending_filled",
                "pending_cancelled"
              ]
            }
          },
          {
            "name": "user",
            "required": false,
            "in": "query",
            "description": "Filter by user address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chain",
            "required": false,
            "in": "query",
            "description": "Filter by source chain",
            "schema": {
              "type": "string",
              "enum": [
                "stellar",
                "ethereum",
                "base",
                "polygon",
                "arbitrum",
                "optimism",
                "avalanche"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of results per page (max 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for the next page of intents",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}": {
      "get": {
        "operationId": "SolversController_getSolver",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      },
      "patch": {
        "operationId": "SolversController_updateSolver",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated solver record"
          },
          "400": {
            "description": "Invalid update body"
          },
          "401": {
            "description": "Missing or invalid signature"
          },
          "404": {
            "description": "Solver not found"
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/stats": {
      "get": {
        "operationId": "SolversController_getSolverStats",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "window",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/slashes": {
      "get": {
        "operationId": "SolversController_getSlashHistory",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/slashes/{slashId}/dispute": {
      "post": {
        "operationId": "SolversController_submitDispute",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slashId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/slashes/{slashId}/dispute/resolve": {
      "post": {
        "operationId": "SolversController_resolveDispute",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slashId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/deregister": {
      "post": {
        "operationId": "SolversController_deregisterSolver",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/deactivate": {
      "post": {
        "operationId": "SolversController_deactivate",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
    "/solvers/{address}/reactivate": {
      "post": {
        "operationId": "SolversController_reactivate",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "solvers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vortex Backend",
    "description": "Intent relay API + WebSocket feed for Vortex Protocol",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.vortex.trade",
      "description": "Production"
    }
  ],
  "components": {
    "schemas": {
...
      "RegisterSolverDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Solver's Stellar address",
            "maxLength": 56
          },
          "name": {
            "type": "string",
            "description": "Solver's display name",
            "maxLength": 64
          },
          "bondAmount": {
            "type": "string",
            "description": "Bond amount as a non-negative integer string (in USDC base units)"
          },
          "avgFillTime": {
            "type": "number",
            "description": "Average fill time in seconds"
          },
          "supportedChains": {
            "type": "array",
            "description": "Chains this solver supports",
            "items": {
              "type": "string",
              "enum": [
                "stellar",
                "ethereum",
                "base",
                "polygon",
                "arbitrum",
                "optimism",
                "avalanche"
              ]
            }
          },
          "supportedTokens": {
            "description": "Token symbols this solver supports",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "proofSignature": {
            "type": "string",
            "description": "Proof-of-control signature for the advertised solver address",
            "maxLength": 88
          }
        },
        "required": [
          "address",
          "name",
          "bondAmount",
          "avgFillTime",
          "supportedChains",
          "supportedTokens",
          "proofSignature"
        ]
      },
      "UpdateSolverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New display name"
          },
          "supportedChains": {
            "type": "array",
            "description": "Replacement list of chains this solver supports",
            "items": {
              "type": "string",
              "enum": [
                "stellar",
                "ethereum",
                "base",
                "polygon",
                "arbitrum",
                "optimism",
                "avalanche"
              ]
            }
          },
          "supportedTokens": {
            "description": "Replacement list of supported token symbols",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "avgFillTime": {
            "type": "number",
            "description": "Updated average fill time in seconds"
          },
          "signature": {
            "type": "string",
            "description": "Base64-encoded Ed25519 signature of the message \"update-solver:<address>\" produced by the solver's private key, proving control of :address"
          }
        },
        "required": [
          "signature"
        ]
      },
      "UpdateSolverStatusDto": {
        "type": "object",
        "properties": {
          "signature": {
            "type": "string",
            "description": "Proof-of-control signature for the solver status update",
            "maxLength": 88
          }
        },
        "required": [
          "signature"
        ]
      },
...
    }
  }
}